+++ NEWS.md
@@ -4,6 +4,7 @@
 * removed curvature for edges and added manual alpha tuning
 * switched from `geom_edge_arc` to `geom_edge_fan`
 * changed some default colors
+* increased max node size
 
 # snahelper 0.1.0